+++ examples/cfd/darcy_transolver/train_transolver_darcy_fix.py
@@ -27,7 +27,8 @@ from physicsnemo.models.transolver import Transolver
 from physicsnemo.distributed import DistributedManager
 from physicsnemo.utils import StaticCaptureTraining, StaticCaptureEvaluateNoGrad
 from physicsnemo.launch.utils import load_checkpoint, save_checkpoint
-from physicsnemo.launch.logging import PythonLogger, LaunchLogger, initialize_mlflow
+from physicsnemo.launch.logging import PythonLogger, LaunchLogger
+from physicsnemo.launch.logging.mlflow import initialize_mlflow
 
 from darcy_datapipe_fix import Darcy2D_fix
 from validator_fix import GridValidator