+++ examples/weather/corrdiff/datasets/cwb.py
@@ -25,9 +25,9 @@ from hydra.utils import to_absolute_path
 import numpy as np
 import zarr
 
-from .base import ChannelMetadata, DownscalingDataset
-from .img_utils import reshape_fields
-from .norm import denormalize, normalize
+from datasets.base import ChannelMetadata, DownscalingDataset
+from datasets.img_utils import reshape_fields
+from datasets.norm import denormalize, normalize
 
 logger = logging.getLogger(__file__)