+++ examples/weather/corrdiff/datasets/gefs_hrrr.py
@@ -22,29 +22,26 @@ import glob
 import logging
 import os
 from typing import Iterable, Tuple, Union
-import cv2
-import s3fs
-import cftime
-import dask
-import numpy as np
-import torch
-from torch.utils.data import DataLoader
-from torch.utils.data.distributed import DistributedSampler
-import xarray as xr
-from physicsnemo.distributed import DistributedManager
-from typing import List, Tuple, Union
-from .base import ChannelMetadata, DownscalingDataset
-import json
 import copy
 
-"""
-TO DO LIST:
-missing samples file
-mean and std
-"""
+import cftime
+import dask
+import json
+import numpy as np
+import torch
+import xarray as xr
+import cv2
 
+from physicsnemo.distributed import DistributedManager
 
-import nvtx
+from datasets.base import ChannelMetadata, DownscalingDataset
+
+# """
+# TO DO LIST:
+# missing samples file
+# mean and std
+# """
+
 
 hrrr_stats_channels = [
     "u10m",