+++ examples/weather/corrdiff/generate.py
@@ -14,26 +14,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import contextlib
+from concurrent.futures import ThreadPoolExecutor
+from functools import partial
+
 import hydra
 from omegaconf import OmegaConf, DictConfig
+from hydra.utils import to_absolute_path
 import torch
 import torch._dynamo
-import nvtx
+from torch.distributed import gather
 import numpy as np
+import nvtx
 import netCDF4 as nc
-import contextlib
 
 from physicsnemo.distributed import DistributedManager
 from physicsnemo.launch.logging import PythonLogger, RankZeroLoggingWrapper
 from physicsnemo.utils.patching import GridPatching2D
 from physicsnemo import Module
-
-from concurrent.futures import ThreadPoolExecutor
-from functools import partial
-from torch.distributed import gather
-
-
-from hydra.utils import to_absolute_path
 from physicsnemo.utils.generative import deterministic_sampler, stochastic_sampler
 from physicsnemo.utils.corrdiff import (
     NetCDFWriter,
@@ -41,7 +39,6 @@ from physicsnemo.utils.corrdiff import (
     regression_step,
     diffusion_step,
 )
-
 
 from helpers.generate_helpers import (
     get_dataset_and_sampler,