+++ examples/weather/corrdiff/inference/plot_multiple_samples.py
@@ -14,8 +14,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import argparse
 import joblib
+import argparse
+
 import matplotlib.pyplot as plt
 import xarray