+++ examples/weather/corrdiff/inference/plot_single_sample.py
@@ -16,6 +16,7 @@
 
 import os
 import argparse
+
 import cftime
 import matplotlib.pyplot as plt
 import netCDF4 as nc