+++ CHANGELOG.md
@@ -38,6 +38,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   amortizing regression costs
 - Explicit handling of Warp device for ball query and sdf
 - Merged SongUNetPosLtEmb with SongUNetPosEmb, add support for batch>1
+- Add lead time embedding support for `positional_embedding_selector`. Enable  
+arbitrary positioning of probabilistic variables
+- Enable lead time aware regression without CE loss
 
 ### Deprecated